error handling

+++ server/Application/Duties/Delete.cs
@@ -1,3 +1,4 @@
+using Application.Core;
 using Domain;
 using MediatR;
 using Persistence;
@@ -6,12 +7,12 @@ namespace Application.Duties;
 
 public class Delete
 {
-    public class Command : IRequest
+    public class Command : IRequest<Result<Unit>>
     {
         public Guid Id { get; set; }
     }
 
-    public class Handler : IRequestHandler<Command>
+    public class Handler : IRequestHandler<Command, Result<Unit>>
     {
         private readonly DataContext _context;
 
@@ -20,15 +21,25 @@ public class Delete
             _context = context;
         }
 
-        public async Task<Unit> Handle(Command request, CancellationToken cancellationToken)
+        public async Task<Result<Unit>> Handle(Command request, CancellationToken cancellationToken)
         {
             Duty duty = await _context.Duties.FindAsync(request.Id);
 
+            if (duty == null)
+            {
+                return null;
+            }
+
             _context.Duties.Remove(duty);
 
-            await _context.SaveChangesAsync();
+            int result = await _context.SaveChangesAsync();
 
-            return Unit.Value;
+            if (result <= 0)
+            {
+                return Result<Unit>.Failure("Failed to delete the duty");
+            }
+
+            return Result<Unit>.Success(Unit.Value);
         }
     }
 }
